Apply #5059: route non-contiguous host buffers (stride int64) [test build]

+++ pjrt_implementation/src/api/loaded_executable_instance.cc
@@ -262,10 +262,9 @@ tt_pjrt_status LoadedExecutableInstance::createDefaultOutputBuffers(
 
     // Row-major strides: last stride is 1, each preceding stride is the
     // product of all following dimension sizes.
-    std::vector<std::uint32_t> strides(output_shape->size());
+    std::vector<std::int64_t> strides(output_shape->size());
     std::exclusive_scan(output_shape->rbegin(), output_shape->rend(),
-                        strides.rbegin(), std::uint32_t(1),
-                        std::multiplies<>());
+                        strides.rbegin(), std::int64_t(1), std::multiplies<>());
 
     // Only create host tensor when not in compile-only mode
     std::optional<tt::runtime::Tensor> host_tensor;
